+++ src/api/markers.ts
@@ -19,7 +19,7 @@ export async function getNotificationMarker(
   })
   const data = await response.json()
   if (response.status !== 200) {
-    throw new Error(data.message)
+    throw new Error(data.error_description)
   }
   return data.notifications
 }
@@ -36,7 +36,7 @@ export async function updateNotificationMarker(
   })
   const data = await response.json()
   if (response.status !== 200) {
-    throw new Error(data.message)
+    throw new Error(data.error_description)
   }
   return data.notifications
 }