+++ src/api/notifications.ts
@@ -25,7 +25,7 @@ export async function getNotifications(
   })
   const data = await response.json()
   if (response.status !== 200) {
-    throw new Error(data.message)
+    throw new Error(data.error_description)
   }
   return data
 }