+++ src/api/settings.ts
@@ -14,7 +14,7 @@ export async function changePassword(
   })
   const data = await response.json()
   if (response.status !== 200) {
-    throw new Error(data.message)
+    throw new Error(data.error_description)
   } else {
     return data
   }
@@ -65,7 +65,7 @@ export async function importFollows(
   })
   if (response.status !== 204) {
     const data = await response.json()
-    throw new Error(data.message)
+    throw new Error(data.error_description)
   }
 }
 
@@ -82,7 +82,7 @@ export async function moveFollowers(
   })
   const data = await response.json()
   if (response.status !== 200) {
-    throw new Error(data.message)
+    throw new Error(data.error_description)
   } else {
     return data
   }