+++ src/api/subscriptions-ethereum.ts
@@ -47,7 +47,7 @@ export async function getSubscriptionAuthorization(
   })
   const data = await response.json()
   if (response.status !== 200) {
-    throw new Error(data.message)
+    throw new Error(data.error_description)
   } else {
     return data
   }