+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Use `/feeds/users/{username}` path for user's Atom feed.
+- Read error messages from `error_description` field instead of `message`.
 
 ## [1.14.0] - 2023-02-22